Bulk update

+++ cli.js
@@ -8,9 +8,10 @@ const rimraf = require("rimraf");
 const basedir = process.cwd();
 msm.walkAll(thisdir => {
   console.log("Hellot here");
-  yarnif.exec("lerna", ["bootstrap"]);
-  yarnif.exec("lerna", ["run", "build"]);
-  yarnif.exec("lerna", ["run", "link"]);
+  yarnif.install();
+  yarnif.exec(["lerna", "bootstrap"]);
+  yarnif.exec(["lerna", "run", "build"]);
+  yarnif.exec(["lerna", "run", "link"]);
   const lernapath = fs.readFileSync("lerna.json");
   const lernapackage = JSON.parse(lernapath);
   if (lernapackage) {
@@ -18,13 +19,15 @@ msm.walkAll(thisdir => {
     const packages = lernapackage.packages || [];
     packages.forEach(g => {
       glob.sync(g).forEach(packagepath => {
-        const packageobj = require(Path.join(packagepath, "package.json"));
+        const fullpackagepath = Path.join(thisdir, packagepath);
+        const packageobj = require(Path.join(fullpackagepath, "package.json"));
         const name = packageobj.name;
         const targetdir = Path.join(basedir, "node_modules", name);
-        fs.mkdir(Path.dirname(targetdir));
-        fs.unlinkSync(targetdir);
+        const parent = Path.dirname(targetdir);
+        if (!fs.existsSync(parent)) fs.mkdirSync(parent);
+        if (fs.existsSync(targetdir)) fs.unlinkSync(targetdir);
         rimraf.sync(targetdir);
-        fs.symlinkSync(packagepath, targetdir);
+        fs.symlinkSync(fullpackagepath, targetdir);
       });
     });
     //Now make my symlink
